Conflicts:
	metagenomics/urls.pyc

+++ contig_retrieval_tool/retrieval_pipeline.pl
@@ -1,28 +1,58 @@
 #!/usr/bin/perl
 use warnings;
 use strict;
+use Cwd qw/abs_path chdir/;
+my $cwd = abs_path('retrieval_pipeline.pl');
+$cwd =~ s/\/retrieval_pipeline.pl//;
+print "HERE $cwd\n";
+chdir("$cwd");
 
-if (scalar(@ARGV == 3)) {
-    my ($f_file, $r_file, $database) = @ARGV;
-    my $pid = `perl endtag_setup.pl $f_file $r_file $database`;
-    $pid = `perl endtag_match.pl $pid`;
-=pod
-    my $ping = `ping -c 1 orca.sharcnet.ca`;
-    if ($ping) {
-        print "Ping $ping\n";
-    }else {
-        print "Ping 1 FAILED\n";
+my $server = 'saw';
+my $filename = 'mg_temp';
+if (scalar(@ARGV) == 3) {
+    open(my $log, ">>", "log.txt");
+    if (setup() == 1) {
+        my ($f_file, $r_file, $database) = @ARGV;
+        print $log "********************************\nExecution of Contig Retrieval Pipeline\n********************************\n";
+        print $log "Executing Endtag Setup\n";
+        print "Prior: $f_file $r_file $database $cwd\n";
+        my $pid = `perl endtag_setup.pl $f_file $r_file $database $cwd`;
+        
+        print $log "Matching the contigs to the cosmids\n";
+        print "After $cwd\n";
+        $pid = `perl endtag_match.pl $pid $cwd`;
+        print $log "Contigs Retrieval Complete\n";     
+
+        system('rm -rf testdb*');
+        #cleanUp();
     }
-    
-=cut
-    $pid = `perl endtag_orf.pl $pid`;
-    $pid = `perl endtag_graphics.pl $pid`;
-    print $pid, "\n";
-    #$pid = `perl endtag_annot.pl $pid`;
-    
-    system('rm -rf testdb*');
+
 } else {
     print help();
+}
+
+
+sub cleanUp{
+    system('rm data.lib');
+    system('rm -rf temp/');
+    system('rm .message');
+}
+
+sub setup{
+    my $ret = 0;
+    (system("mkdir temp")) unless (-d "temp");    
+    (system("mkdir -p temp/tmp")) unless (-d "temp/tmp");
+    (system("mkdir -p tmp/out")) unless (-d "tmp/out");
+    (system("mkdir -p temp/data/blast")) unless (-d "temp/data");
+    (system("mkdir -p tmp/img")) unless (-d "tmp/img");
+    (system("mkdir temp/orf_seq")) unless (-d "temp/orf_seq");
+    (system("mkdir -p temp/predicts/blast")) unless (-d "temp/predicts");
+    (system("mkdir temp/storage")) unless (-d "temp/storage");
+    (system("mkdir temp/genbank")) unless (-d "temp/genbank");
+    if ((-d "temp/tmp") && (-d "temp/data") && (-d "tmp/img") && (-d "tmp/out") && (-d "temp/genbank") && (-d "temp/orf_seq") && (-d "temp/predicts") && (-d "temp/storage")){
+        $ret = 1;
+    }
+    return $ret;
 }
 
 sub help{
@@ -52,11 +82,3 @@ HELPFILE
     return $help;
 }
 
-__END__
-#!/bin/bash
-sqsub -r 15m -n 16 -q mpi --mpp=1G -o ofile%J mpirun -np 16 --machinefile $PBS_NODEFILE mpiblast -d drosoph.nt -i drosoph.fa -p blastn -o drosoph.out --use-parallel-write --use-virtual-frags
-/opt/sharcnet/openmpi/1.6.2/intel/bin/mpirun -np 16 --machinefile $PBS_NODEFILE mpiblast -d drosoph.nt -i drosoph2.fa -p blastn -o drosoph2.out --use-parallel-write --use-virtual-frags
-/opt/sharcnet/openmpi/1.6.2/intel/bin/mpirun -np 16 --machinefile $PBS_NODEFILE mpiblast -d drosoph.nt -i drosoph3.fa -p blastn -o drosoph3.out --use-parallel-write --use-virtual-frags
-sqsub -r 15m -n 16 -q mpi --mpp=1G -o ofile%J ./run.sh
-sqjobs | egrep -o '[0-9]+?\s*?mpi\s*?R' | egrep -o '[0-9]+'
-sqjobs | egrep -o '3670361+?\s*?mpi\s*?.' | egrep -o '[A-Z]'
